Add a toggle to enable/disable cookies in the browser
If Dillo is compiled with -DDISABLE_COOKIES then the menu
entry is not compiled. If the user decides to include cookies
then a menu item will show up, giving the user to disable
the cookies for the current session.
(The state is not saved anywhere)

Two functions have been added to cookies.c to check and set
the "disabled" variable (thus enabling/blocking cookies)

menu.cc looks a bit chaotic, but should be fine (?)