Make CTRL+A select text in the current field
Change the default behavior of CTRL+A from moving cursor to the
beginning of the line, to selecting the whole text in the current.
This is more intuivite for casual users and goes in pair with
CTLR+X/C/V shortcuts.

Modified files:
dw/fltkui.cc <- shortcut inside the webpage
src/dialog.cc <- shortcut in a dialog (not tested)
src/fidnbar.cc <- shortcut in Search field
src/ui.cc <- shortcut in address bar